layer list: split components into modules.
