+++ src/renderer/components/layerlist/Search.js
@@ -1,0 +1,50 @@
+import { ipcRenderer } from 'electron'
+import React from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import { InputBase } from '@material-ui/core'
+
+const useStyles = makeStyles((theme) => ({
+  search: {
+    paddingLeft: '8px',
+    paddingRight: '8px',
+    paddingTop: theme.spacing(1),
+    paddingBottom: theme.spacing(1),
+    fontSize: '120%',
+    userSelect: 'none'
+  }
+}))
+
+export const Search = (/* props */) => {
+  const classes = useStyles()
+
+  // Search: Prevent undo/redo when not focused.
+  const [readOnly, setReadOnly] = React.useState(false)
+  const ref = React.useRef()
+  const selectAll = React.useCallback(() => {
+    const input = ref.current
+    input.setSelectionRange(0, input.value.length)
+  }, [])
+
+  const handleBlur = () => {
+    ipcRenderer.removeListener('IPC_EDIT_SELECT_ALL', selectAll)
+    setReadOnly(true)
+  }
+
+  const handleFocus = () => {
+    ipcRenderer.on('IPC_EDIT_SELECT_ALL', selectAll)
+    setReadOnly(false)
+  }
+
+  return (
+    <InputBase
+      className={classes.search}
+      placeholder={'Search...'}
+      autoFocus
+      inputRef={ref}
+      readOnly={readOnly}
+      onFocus={handleFocus}
+      onBlur={handleBlur}
+      disabled
+    />
+  )
+}
